requirements.txt
pillow>=10.0.0
pystray>=0.19.4
watchdog>=3.0.0
psutil>=5.9.0
pywin32>=306; platform_system == "Windows"